+++ src/consumer/versions-ops/merge-version/index.js
@@ -1,5 +1,12 @@
 // @flow
 
 export { default as threeWayMerge } from './three-way-merge';
-export { mergeVersion, FileStatus, getMergeStrategyInteractive, getMergeStrategy, MergeOptions } from './merge-version';
+export {
+  mergeVersion,
+  FileStatus,
+  getMergeStrategyInteractive,
+  getMergeStrategy,
+  MergeOptions,
+  filesStatusWithoutSharedDir
+} from './merge-version';
 export type { MergeStrategy, ApplyVersionResults, ApplyVersionResult } from './merge-version';