+++ src/scope/scope.js
@@ -18,7 +18,15 @@ import types from './object-registrar';
 import { propogateUntil, currentDirName, pathHas, first, readFile, splitBy, pathNormalizeToLinux } from '../utils';
 import { BIT_HIDDEN_DIR, LATEST, OBJECTS_DIR, BITS_DIRNAME, BIT_VERSION, DEFAULT_BIT_VERSION } from '../constants';
 import { ScopeJson, getPath as getScopeJsonPath } from './scope-json';
-import { ScopeNotFound, ComponentNotFound, DependencyNotFound, CyclicDependencies } from './exceptions';
+import {
+  ScopeNotFound,
+  ComponentNotFound,
+  ResolutionException,
+  DependencyNotFound,
+  CyclicDependencies,
+  MergeConflictOnRemote,
+  MergeConflict
+} from './exceptions';
 import IsolatedEnvironment from '../environment';
 import { RemoteScopeNotFound, PermissionDenied } from './network/exceptions';
 import { Tmp } from './repositories';
@@ -29,6 +37,7 @@ import { Repository, Ref, BitObject } from './objects';
 import ComponentWithDependencies from './component-dependencies';
 import VersionDependencies from './version-dependencies';
 import SourcesRepository from './repositories/sources';
+import type { ComponentTree } from './repositories/sources';
 import Consumer from '../consumer/consumer';
 import { index } from '../search/indexer';
 import loader from '../cli/loader';
@@ -589,6 +598,29 @@ export default class Scope {
     componentsObjects.component.scope = remoteScope;
   }
 
+  async _mergeObjects(manyObjects: ComponentTree[]) {
+    const mergeResults = await Promise.all(
+      manyObjects.map(async (objects) => {
+        try {
+          const result = await this.sources.merge(objects, true, false);
+          return result;
+        } catch (err) {
+          if (err instanceof MergeConflict) {
+            return err; // don't throw. instead, get all components with merge-conflicts
+          }
+          throw err;
+        }
+      })
+    );
+    const componentsWithConflicts = mergeResults.filter(result => result instanceof MergeConflict);
+    if (componentsWithConflicts.length) {
+      const idsAndVersions = componentsWithConflicts.map(c => ({ id: c.id, versions: c.versions }));
+      // sort to have a consistent error message
+      const idsAndVersionsSorted = R.sortBy(R.prop('id'), idsAndVersions);
+      throw new MergeConflictOnRemote(idsAndVersionsSorted);
+    }
+  }
+
   /**
    * @TODO there is no real difference between bare scope and a working directory scope - let's adjust terminology to avoid confusions in the future
    * saves a component into the objects directory of the remote scope, then, resolves its
@@ -602,7 +634,7 @@ export default class Scope {
       `exportManyBareScope: Going to save ${componentsObjects.length} components`
     );
     const manyObjects = componentsObjects.map(componentObjects => componentObjects.toObjects(this.objects));
-    await Promise.all(manyObjects.map(objects => this.sources.merge(objects, true, false)));
+    await this._mergeObjects(manyObjects);
     const manyCompVersions = await Promise.all(
       manyObjects.map(objects => objects.component.toComponentVersion(LATEST))
     );
@@ -924,13 +956,18 @@ export default class Scope {
    * Remove components from scope
    * @force Boolean  - remove component from scope even if other components use it
    */
-  async removeMany(bitIds: BitIds, force: boolean, removeSameOrigin: boolean = false): Promise<RemovedObjects> {
+  async removeMany(
+    bitIds: BitIds,
+    force: boolean,
+    removeSameOrigin: boolean = false,
+    consumer?: Consumer
+  ): Promise<RemovedObjects> {
     logger.debug(`scope.removeMany ${bitIds} with force flag: ${force.toString()}`);
     Analytics.addBreadCrumb(
       'removeMany',
       `scope.removeMany ${Analytics.hashData(bitIds)} with force flag: ${force.toString()}`
     );
-    const removeComponents = new RemoveModelComponents(this, bitIds, force, removeSameOrigin);
+    const removeComponents = new RemoveModelComponents(this, bitIds, force, removeSameOrigin, consumer);
     return removeComponents.remove();
   }