+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- improve merge-conflict error on export to show all components with conflicts
+- fix `bit remove` to not delete dependencies when they were imported directly
+
 ## [0.12.12] - 2018-04-29
 
 ### New